backend/microservices/automatic_translation/requirements.txt
fastapi
uvicorn
openai
pillow
python-dotenv
python-multipart
